+++ src/components/HelpModal/HelpModal.tsx
@@ -1,6 +1,7 @@
 import React, { ReactNode } from 'react';
-import { Box, Card, CardContent, CardHeader, IconButton, Modal } from '@mui/material';
+import { Box, Card, CardContent, CardHeader, Modal } from '@mui/material';
 import HighlightOffIcon from '@mui/icons-material/HighlightOff';
+import { OurIconButton } from '../IconButton/IconButton';
 
 export type Help = {
   content: Object;
@@ -8,12 +9,14 @@ export type Help = {
   showHelp: Boolean;
 };
 export interface HelpModalProps {
+  ariaDescription?: string;
   component?: boolean;
   help: Help;
   helpToggle: Function;
+  toolTip?: string;
 }
 
-export function HelpModal({ component = false, help, helpToggle }: HelpModalProps): JSX.Element {
+export function HelpModal({ ariaDescription = "Help Modal", component = false, help, helpToggle, toolTip = "Help close button" }: HelpModalProps): JSX.Element {
   function closeModal() {
     helpToggle();
   }
@@ -24,15 +27,13 @@ export function HelpModal({ component = false, help, helpToggle }: HelpModalProp
         <Card variant="outlined" sx={{ height: '90vh', width: '90vw' }}>
           <CardHeader
             action={
-              <IconButton
-                aria-label="DUMMY"
-                sx={{ '&:hover': { backgroundColor: 'primary.dark' }, ml: 1 }}
+              <OurIconButton
+                ariaDescription={ariaDescription}
+                icon={<HighlightOffIcon />}
                 onClick={() => closeModal()}
-                color="inherit"
-                data-testid="closeHelpModalBtn"
-              >
-                <HighlightOffIcon />
-              </IconButton>
+                testId={'closeHelpModalBtn'}
+                toolTip={toolTip}
+              />
             }
           />
           <CardContent sx={{ height: '90vh', width: '90vw' }}>