+++ src/components/version/index.ts
@@ -1,1 +1,1 @@
-export const GUI_COMPONENTS_VERSION = '2.1.02';
+export const GUI_COMPONENTS_VERSION = '2.1.03';