+++ types/components/CopyrightModal/CopyrightModal.d.ts
@@ -1,6 +1,9 @@
 export interface CopyrightModalProps {
-  show: boolean;
-  copyrightFunc: Function;
+    ariaDescription?: string;
+    copyrightFunc: Function;
+    show: boolean;
+    title?: string;
+    toolTip?: string;
 }
-export declare function CopyrightModal({ show, copyrightFunc }: CopyrightModalProps): JSX.Element;
+export declare function CopyrightModal({ ariaDescription, copyrightFunc, show, title, toolTip }: CopyrightModalProps): JSX.Element;
 export default CopyrightModal;